+++ releasenotes.md
@@ -2,6 +2,21 @@
 
 ## 进展
 
+## Version 1.6.0 2020-10-13 20:22:28
+
+新功能：
+
+- [关键定位功能](ofdrw-reader/src/test/java/org/ofdrw/reader/keyword/KeywordExtractorTest.java)
+- [骑缝章的自由定位](ofdrw-sign/src/test/java/org/ofdrw/sign/stamppos/RidingStampPosTest.java)
+- [不同版本电子签章兼容解析](ofdrw-gm/src/test/java/org/ofdrw/gm/ses/parse/VersionParserTest.java)
+
+修复:
+
+- 修复了 CustomDatas 在解析CustomData 序列时候错误的关键字使用。
+- 修复了V4版本电子签章可选参数tag0的错误。
+- 兼容非标准推荐的签名ID，如"sN"、"N"类型的解析。
+- 支持了坐标偏移中厂商定义的`g`参数偏移量。
+
 ## Version 1.5.6 2020-9-21 23:14:58
 
 - 修复了无法解析OFD内含有中文目录的文件错误。